Refactor: rename text_processing to nlp "text_processing" is clearer, but the import statement is longer. "nlp" is shorter and makes it clear that it is natural language processing.
