+++ style_bert_vits2/nlp/english/bert_feature.py
@@ -5,8 +5,7 @@ import torch
 from transformers import PreTrainedModel
 
 from style_bert_vits2.constants import Languages
-from style_bert_vits2.text_processing import bert_models
-from style_bert_vits2.text_processing.japanese.g2p import text_to_sep_kata
+from style_bert_vits2.nlp import bert_models
 
 
 __models: dict[torch.device | str, PreTrainedModel] = {}
@@ -20,10 +19,10 @@ def extract_bert_feature(
     assist_text_weight: float = 0.7,
 ) -> torch.Tensor:
     """
-    日本語のテキストから BERT の特徴量を抽出する
+    英語のテキストから BERT の特徴量を抽出する
 
     Args:
-        text (str): 日本語のテキスト
+        text (str): 英語のテキスト
         word2ph (list[int]): 元のテキストの各文字に音素が何個割り当てられるかを表すリスト
         device (torch.device | str): 推論に利用するデバイス
         assist_text (Optional[str], optional): 補助テキスト (デフォルト: None)
@@ -33,12 +32,6 @@ def extract_bert_feature(
         torch.Tensor: BERT の特徴量
     """
 
-    # 各単語が何文字かを作る `word2ph` を使う必要があるので、読めない文字は必ず無視する
-    # でないと `word2ph` の結果とテキストの文字数結果が整合性が取れない
-    text = "".join(text_to_sep_kata(text, raise_yomi_error=False)[0])
-
-    if assist_text:
-        assist_text = "".join(text_to_sep_kata(assist_text, raise_yomi_error=False)[0])
     if (
         sys.platform == "darwin"
         and torch.backends.mps.is_available()
@@ -50,11 +43,11 @@ def extract_bert_feature(
     if device == "cuda" and not torch.cuda.is_available():
         device = "cpu"
     if device not in __models.keys():
-        __models[device] = bert_models.load_model(Languages.JP).to(device)  # type: ignore
+        __models[device] = bert_models.load_model(Languages.EN).to(device)  # type: ignore
 
     style_res_mean = None
     with torch.no_grad():
-        tokenizer = bert_models.load_tokenizer(Languages.JP)
+        tokenizer = bert_models.load_tokenizer(Languages.EN)
         inputs = tokenizer(text, return_tensors="pt")
         for i in inputs:
             inputs[i] = inputs[i].to(device)  # type: ignore
@@ -68,7 +61,7 @@ def extract_bert_feature(
             style_res = torch.cat(style_res["hidden_states"][-3:-2], -1)[0].cpu()
             style_res_mean = style_res.mean(0)
 
-    assert len(word2ph) == len(text) + 2, text
+    assert len(word2ph) == res.shape[0], (text, res.shape[0], len(word2ph))
     word2phone = word2ph
     phone_level_feature = []
     for i in range(len(word2phone)):
