+++ style_bert_vits2/nlp/japanese/__init__.py
@@ -1,0 +1,2 @@
+from style_bert_vits2.nlp.japanese.g2p import g2p  # noqa: F401
+from style_bert_vits2.nlp.japanese.normalizer import normalize_text  # noqa: F401
