+++ style_bert_vits2/nlp/japanese/g2p.py
@@ -2,11 +2,11 @@ import re
 
 from style_bert_vits2.constants import Languages
 from style_bert_vits2.logging import logger
-from style_bert_vits2.text_processing import bert_models
-from style_bert_vits2.text_processing.japanese import pyopenjtalk_worker as pyopenjtalk
-from style_bert_vits2.text_processing.japanese.mora_list import MORA_KATA_TO_MORA_PHONEMES
-from style_bert_vits2.text_processing.japanese.normalizer import replace_punctuation
-from style_bert_vits2.text_processing.symbols import PUNCTUATIONS
+from style_bert_vits2.nlp import bert_models
+from style_bert_vits2.nlp.japanese import pyopenjtalk_worker as pyopenjtalk
+from style_bert_vits2.nlp.japanese.mora_list import MORA_KATA_TO_MORA_PHONEMES
+from style_bert_vits2.nlp.japanese.normalizer import replace_punctuation
+from style_bert_vits2.nlp.symbols import PUNCTUATIONS
 
 pyopenjtalk.initialize()
