+++ style_bert_vits2/nlp/japanese/normalizer.py
@@ -2,7 +2,7 @@ import re
 import unicodedata
 from num2words import num2words
 
-from style_bert_vits2.text_processing.symbols import PUNCTUATIONS
+from style_bert_vits2.nlp.symbols import PUNCTUATIONS
 
 
 def normalize_text(text: str) -> str:
