+++ style_bert_vits2/nlp/japanese/pyopenjtalk_worker/worker_common.py
