+++ style_bert_vits2/nlp/japanese/user_dict/__init__.py
@@ -16,9 +16,9 @@ import numpy as np
 from fastapi import HTTPException
 
 from style_bert_vits2.constants import DEFAULT_USER_DICT_DIR
-from style_bert_vits2.text_processing.japanese import pyopenjtalk_worker as pyopenjtalk
-from style_bert_vits2.text_processing.japanese.user_dict.word_model import UserDictWord, WordTypes
-from style_bert_vits2.text_processing.japanese.user_dict.part_of_speech_data import MAX_PRIORITY, MIN_PRIORITY, part_of_speech_data
+from style_bert_vits2.nlp.japanese import pyopenjtalk_worker as pyopenjtalk
+from style_bert_vits2.nlp.japanese.user_dict.word_model import UserDictWord, WordTypes
+from style_bert_vits2.nlp.japanese.user_dict.part_of_speech_data import MAX_PRIORITY, MIN_PRIORITY, part_of_speech_data
 
 pyopenjtalk.initialize()
