--- style_bert_vits2/text_processing/japanese/__init__.py
@@ -1,2 +1,0 @@
-from style_bert_vits2.text_processing.japanese.g2p import g2p  # noqa: F401
-from style_bert_vits2.text_processing.japanese.normalizer import normalize_text  # noqa: F401
