+++ train_ms_jp_extra.py
@@ -32,7 +32,7 @@ from style_bert_vits2.models.models_jp_extra import (
     SynthesizerTrn,
     WavLMDiscriminator,
 )
-from style_bert_vits2.text_processing.symbols import SYMBOLS
+from style_bert_vits2.nlp.symbols import SYMBOLS
 from style_bert_vits2.utils.stdout_wrapper import SAFE_STDOUT
 
 torch.backends.cuda.matmul.allow_tf32 = True
